refactor(std-net): extract probe substrate and add net_capabilities()
Restructure std/net into a directory module to give probe tooling a
shared foundation ahead of traceroute:

- net/probe.rs: ProbeFailure { Target, Backend } carries the
  target-vs-backend failure distinction as a type through every probe
  layer; classification happens once at the io boundary instead of
  string-matching composed error messages. Shared probe_attempt_budget
  divides one global deadline across attempts and intervals.
- net/icmp.rs: ICMP substrate (datagram-first/raw-fallback sockets,
  echo packet construction, checksums, reply/error parsing, capability
  detection) separated from the ping driver so traceroute can reuse it.
- net_capabilities(): reports available probe paths (datagram/raw ICMP
  per family, tcp) without sending traffic, with typechecker signature,
  docs, and a CI-safe integration test.
- build.rs: scan src/stdlib/ recursively so directory modules keep
  their @ntnt docs validated and generated (the flat scan silently
  dropped std/net docs after the module split).
- dd-046: record PR #119/#118 in the status dashboard, promote
  traceroute from deferred to a planned Phase 2 with its remaining
  capability checklist.

User-visible behavior of ping() is unchanged; error message formats are
preserved.
